+++ runtime/src/main/java/com/dylibso/chicory/runtime/TableInstance.java
@@ -56,7 +56,7 @@ public class TableInstance {
         return this.refs[index];
     }
 
-    public void setRef(int index, long value, Instance instance) {
+    public void setRef(int index, int value, Instance instance) {
         if (index < 0 || index >= this.refs.length || index >= this.instances.length) {
             throw new UninstantiableException("out of bounds table access");
         }
